+++ client/src/stories/DetailsPage.stories.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { DetailsPage } from "./DetailsPage";
+import { DetailsPage } from "../pages/DetailsPage";
 
 export default {
   title: "Pages/DetailsPage",
